Mon, Jun 26, 2023, 7:02 AM -07:00

+++ filter-more/helpers.c
@@ -70,9 +70,9 @@ void blur(int height, int width, RGBTRIPLE image[height][width])
             // int count = 0;
 
             //for pixels that are not edge cases
-            long red = 0x00;
-            long green = 0x00;
-            long blue = 0x00;
+            int red = 0x00;
+            int green = 0x00;
+            int blue = 0x00;
             if (j > 0 && j < width - 1 && i > 0 && i < height){
                 red += image[i-1][j-1].rgbtRed;
                 green += image[i-1][j-1].rgbtGreen;
